backend/rag-chatbot/requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain_openai==0.0.5
langchain_chroma==0.1.0
langchain_community==0.0.12
langchain_text_splitters==1.1.2
chromadb==0.4.22
openai==1.10.0

# Web framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.9

# Vector operations
numpy==1.26.3
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
gradio==4.7.1
